refactor: split keymap stuff into multiple files
